# api-documentation.yaml
openapi: 3.0.3
info:
  title: AI-HR Recruitment System API
  description: |
    Complete AI-driven recruitment system with resume filtering, AI interview chat, and candidate scoring functionality.

    ## System Workflow
    1. **User Management**: Candidate registration/login, HR login and create other HR accounts
    2. **Job Management**: HR creates and closes job positions
    3. **Application Process**: Candidates apply for positions and submit resumes
    4. **AI Assessment**: AI filters resumes and scores, chats with candidates and scores again
    5. **HR Review**: HR views job applicants, scores, and resumes

    ## Authentication
    Uses RSA-signed JWT Bearer Token for authentication.

  version: 2.0.0
  contact:
    name: AI-HR Team
    email: yifei.liu@tum.de
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development environment

# Security configuration
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RSA-signed JWT token

  # Data model definitions
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        message:
          type: string
          description: Response message
        data:
          oneOf:
            - type: object
            - type: array
            - type: "null"
          description: Response data
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        code:
          type: integer
          description: HTTP status code
      required:
        - success
        - message
        - timestamp
        - code

    # Error response
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time
        code:
          type: integer

    # Pagination information
    PageInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
          description: Current page number (0-based)
        size:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        totalPages:
          type: integer
          description: Total number of pages

    # User-related models
    UserDto:
      type: object
      properties:
        userID:
          type: string
          format: uuid
        fullName:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        role:
          type: string
          enum: [CANDIDATE, HR]
        creationTimestamp:
          type: string
          format: date-time
      required:
        - userID
        - fullName
        - email
        - role

    # Authentication-related models
...
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 6
      required:
        - fullName
        - email
        - password
...
    JobDto:
      type: object
      properties:
        jobID:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        description:
          type: string
        requirements:
          type: string
          description: Job requirements
        status:
          type: string
          enum: [OPEN, CLOSED, DRAFT]
        creationTimestamp:
          type: string
          format: date-time
        closingDate:
          type: string
          format: date
        lastModifiedTimestamp:
          type: string
          format: date-time
        hrCreator:
          $ref: "#/components/schemas/UserDto"
      required:
        - jobID
        - title
        - status
        - hrCreator
    CreateJobRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        requirements:
          type: string
        closingDate:
          type: string
          format: date
      required:
        - title
        - description
        - requirements
    UpdateJobRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        requirements:
          type: string
        closingDate:
          type: string
          format: date
        status:
          type: string
          enum: [OPEN, CLOSED, DRAFT]

    # Application-related models
...
    ChatSessionDto:
      type: object
      properties:
        sessionID:
          type: string
          format: uuid
        applicationID:
          type: string
          format: uuid
        status:
          type: string
          enum: [ACTIVE, COMPLETED, EXPIRED]
        startTimestamp:
          type: string
          format: date-time
        endTimestamp:
          type: string
          format: date-time
        messageCount:
          type: integer
          description: Total number of messages
      required:
        - sessionID
        - applicationID
        - status
...
paths:
...
  /users/hr:
    post:
      operationId: createHRUser
      tags:
        - User Management
      summary: Create HR user
      description: Existing HR users create HR accounts for others
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: HR user created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserDto"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ===================
  # Job Management APIs (Job Service)
  # ===================
  /jobs:
    get:
      operationId: getJobList
      tags:
        - Job Management
      summary: Get job list
      description: Get all open job positions (paginated)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, CLOSED, DRAFT]
      responses:
        "200":
          description: Get successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            type: array
                            items:
                              $ref: "#/components/schemas/JobDto"
                          pageInfo:
                            $ref: "#/components/schemas/PageInfo"

    post:
      operationId: createJob
      tags:
        - Job Management
      summary: Create job position
      description: HR users create new job positions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
            example:
              title: "Senior Java Developer"
              description: "Responsible for backend system development and maintenance"
              requirements: "3+ years Java development experience, familiar with Spring framework"
              closingDate: "2024-02-28"
      responses:
        "201":
          description: Creation successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/JobDto"
        "403":
          description: Insufficient permissions (non-HR user)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{jobId}:
    get:
      operationId: getJobDetails
      tags:
        - Job Management
      summary: Get job details
      description: Get detailed information of a specific job position
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Get successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/JobDto"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateJob
      tags:
        - Job Management
      summary: Update job position
      description: HR users update job position information
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Update successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/JobDto"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{jobId}/close:
    post:
      operationId: closeJob
      tags:
        - Job Management
      summary: Close job position
      description: HR users close specified job position
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Close successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/JobDto"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ===================
  # Application Management APIs (Application Service)
  # ===================
...
  /applications/{applicationId}/chat:
    get:
      operationId: getChatSession
      tags:
        - AI Chat Interview
      summary: Get chat session
      description: Get AI chat session information related to the application
      security:
        - BearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Get successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ChatSessionDto"
        "404":
          description: Chat session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: startChatInterview
      tags:
        - AI Chat Interview
      summary: Start AI chat interview
      description: Start AI chat interview session for specified application
      security:
        - BearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Chat session created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ChatSessionDto"
        "400":
          description: Unable to create chat session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
security:
  - BearerAuth: []