# service-genai/app/routers/chat.py
from typing import List
from uuid import UUID

from fastapi import APIRouter, HTTPException, status
from pydantic import BaseModel

router = APIRouter(
    prefix="/chat",
    tags=["chat"],
    responses={404: {"description": "Not found"}},
)


class Message(BaseModel):
    sender: str  # "AI" or "CANDIDATE"
    content: str


class ChatRequest(BaseModel):
    application_id: UUID
    messages: List[Message]


class ChatResponse(BaseModel):
    response: str
    session_id: UUID


@router.post("/", response_model=ChatResponse)
async def chat_with_ai(request: ChatRequest):
    """
    Process a chat message and generate an AI response.
    
    This endpoint will:
    1. Retrieve relevant data from pgvector and other tables
    2. Process the conversation history
    3. Generate an AI response based on the resume and job requirements
    
    Args:
        request: ChatRequest containing application_id and message history
        
    Returns:
        AI response and session ID
    """
    try:
        # Placeholder for actual implementation
        # In the future, this will:
        # 1. Get relevant data from database
        # 2. Process conversation with AI
        # 3. Store the message in the database

        # Mock response with a dummy session ID
        return ChatResponse(
            response="This is a placeholder response from the AI. The actual implementation will process your resume and job requirements to provide meaningful responses.",
            session_id=request.application_id  # Using application_id as session_id for now
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to process chat: {str(e)}"
        )